chore: move all queries to queries.js
