+++ src/components/app/data/queries/index.js
@@ -1,6 +1,5 @@
 import { queryEnterpriseLearner } from './queries';
 
-export * from './subsidies';
 export * from './queries';
 
 /**
