--- src/components/app/data/queries/subsidies/couponCodes.js
@@ -1,18 +1,0 @@
-import queries from '../queryKeyFactory';
-
-/**
- * Helper function to assist querying with useQuery package
- * queries
- * .enterprise
- * .enterpriseCustomer(enterpriseUuid)
- * ._ctx.subsidies
- * ._ctx.couponCodes
- * @returns
- */
-export default function queryCouponCodes(enterpriseUuid) {
-  return queries
-    .enterprise
-    .enterpriseCustomer(enterpriseUuid)
-    ._ctx.subsidies
-    ._ctx.couponCodes;
-}
