--- src/components/app/data/queries/subsidies/index.js
@@ -1,6 +1,0 @@
-export { default as queryCouponCodes } from './couponCodes';
-export { default as queryEnterpriseLearnerOffers } from './enterpriseOffers';
-export { default as queryRedeemablePolicies } from './policies';
-export { default as querySubscriptions } from './subscriptions';
-
-export * from './browseAndRequest';
