--- src/components/app/data/queries/subsidies/policies.js
@@ -1,20 +1,0 @@
-import queries from '../queryKeyFactory';
-
-/**
- * Helper function to assist querying with useQuery package
- * queries
- * .enterprise
- * .enterpriseCustomer(enterpriseUuid)
- * ._ctx.subsidies
- * ._ctx.policy
- * ._ctx.redeemablePolicies(lmsUserId)
- * @returns
- */
-export default function queryRedeemablePolicies({ enterpriseUuid, lmsUserId }) {
-  return queries
-    .enterprise
-    .enterpriseCustomer(enterpriseUuid)
-    ._ctx.subsidies
-    ._ctx.policy
-    ._ctx.redeemablePolicies(lmsUserId);
-}
